backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
motor
python-dotenv
